Reland "Move is<T> and as<T> logic into IRNode."
This reverts commit da8302ed84dd7a17f053e0b3a730630e69a875b2.

Reason for revert: add cast to placate GCC

Original change's description:
> Revert "Move is<T> and as<T> logic into IRNode."
>
> This reverts commit c7de283e2a252b782793dd5c00d44de62012ccb6.
>
> Reason for revert: red on tree
>
> Original change's description:
> > Move is<T> and as<T> logic into IRNode.
> >
> > Previously, we repeated this logic in four places because the Kind
> > enums lived inside the IRNode subtypes. Moving the Kind enums into
> > IRNode is fine, and lets us centralize these helpers in one place.
> > It also unlocks the ability to work directly with IRNodes, i.e.
> > `myIRNode->is<ReturnStatement>()` is now supported.
> >
> > Reviewed-on: https://skia-review.googlesource.com/c/skia/+/590539
> > Commit-Queue: John Stiles <johnstiles@google.com>
>
> No-Presubmit: true
> No-Tree-Checks: true
> No-Try: true
> Reviewed-on: https://skia-review.googlesource.com/c/skia/+/590745
> Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
> Auto-Submit: John Stiles <johnstiles@google.com>
> Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/590856
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>

+++ src/sksl/codegen/SkSLWGSLCodeGenerator.cpp
@@ -685,7 +685,8 @@ void WGSLCodeGenerator::writeStatement(const Statement& s) {
             this->writeVarDeclaration(s.as<VarDeclaration>());
             break;
         default:
-            SkDEBUGFAILF("unsupported statement (kind: %d) %s", s.kind(), s.description().c_str());
+            SkDEBUGFAILF("unsupported statement (kind: %d) %s",
+                         static_cast<int>(s.kind()), s.description().c_str());
             break;
     }
 }
